Set up a GNU makefile, added utilities, moved files.
The makefile supports running tests incrementally based on changes to
source file contents.  "raco make" does the heavy lifting to track
dependencies.

The utilities are under utils, and include grep-like functions in
racket and power shell.

The files were moved so that we can grow a set of tested files rather
than have a workspace of mixed tested and untested files.